RavenDB-18201 cleanup
